Added a HasCallbackProperties class and a ListCallbackProperty class
The HasCallbackProperties class can be used to add callbacks directly from a class method, including to add a callback to *all* callback properties.

The  ListCallbackProperty class triggers a notification if anything in the list is changed, including if the list contains a class that inherits from HasCallbackProperties which itself has properties that change.